+++ src/rp2040/pico_platform/CMakeLists.txt
@@ -17,6 +17,7 @@ if (NOT TARGET pico_platform)
 
     target_link_libraries(pico_platform_headers INTERFACE hardware_regs)
     pico_mirrored_target_link_libraries(pico_platform INTERFACE
+            pico_platform_common
             pico_platform_compiler
             pico_platform_panic
             pico_platform_sections
