+++ src/rp2_common/hardware_exception/exception.c
@@ -31,12 +31,14 @@ static inline exception_handler_t *get_exception_table(void) {
 #endif
 }
 
+#if !PICO_NO_RAM_VECTOR_TABLE
 static void set_raw_exception_handler_and_restore_interrupts(enum exception_number num, exception_handler_t handler, uint32_t save) {
     // update vtable (vtable_handler may be same or updated depending on cases, but we do it anyway for compactness)
     get_exception_table()[num] = handler;
     __dmb();
     restore_interrupts_from_disabled(save);
 }
+#endif
 
 static inline void check_exception_param(__unused enum exception_number num) {
     invalid_params_if(HARDWARE_EXCEPTION, num < MIN_EXCEPTION_NUM || num > MAX_EXCEPTION_NUM);
@@ -54,10 +56,12 @@ exception_handler_t exception_set_exclusive_handler(enum exception_number num, e
     exception_handler_t current = exception_get_vtable_handler(num);
     hard_assert(handler == current || exception_is_compile_time_default(current));
     set_raw_exception_handler_and_restore_interrupts(num, handler, save);
+    return current;
 #else
+    ((void)num);
+    ((void)handler);
     panic_unsupported();
 #endif
-    return current;
 }
 
 void exception_restore_handler(enum exception_number num, exception_handler_t original_handler) {
@@ -66,6 +70,8 @@ void exception_restore_handler(enum exception_number num, exception_handler_t or
     uint32_t save = save_and_disable_interrupts();
     set_raw_exception_handler_and_restore_interrupts(num, original_handler, save);
 #else
+    ((void)num);
+    ((void)original_handler);
     panic_unsupported();
 #endif
 }
