+++ src/rp2_common/pico_crt0/crt0.S
@@ -23,14 +23,23 @@
 
 pico_default_asm_setup
 
+#ifdef PICO_NO_STORED_VECTOR_TABLE
+#warning PICO_NO_STORED_VECTOR_TABLE is no longer used. PICO_MINIMAL_STORED_VECTOR_TABLE is not identical but usually serves the same purpose
+#endif
+
 .section .vectors, "ax"
 .align 2
 
-.global __vectors, __VECTOR_TABLE
+.global __vectors, __VECTOR_TABLE, __vectors_end
 __VECTOR_TABLE:
 __vectors:
 .word __StackTop
 .word _reset_handler
+
+#if PICO_MINIMAL_STORED_VECTOR_TABLE
+.word isr_invalid   // NMI
+.word isr_invalid   // HardFault
+#else
 .word isr_nmi
 .word isr_hardfault
 .word isr_invalid // Reserved, should never fire
@@ -45,15 +54,9 @@ __vectors:
 .word isr_invalid // Reserved, should never fire
 .word isr_pendsv
 .word isr_systick
-#if PICO_NO_STORED_VECTOR_TABLE && !PICO_NO_FLASH // note in no flash binary, we only have the single RAM vector table anyway
-#if PICO_NO_RAM_VECTOR_TABLE
-#error Can't specify PICO_NO_STORED_VECTOR_TABLE and PICO_NO_RAM_VECTOR_TABLE
-#endif
-// we don't include any IRQ vectors; we will initialize them during runtime_init in the RAM vector table
-#else
 
 .macro if_irq_word num func
-.if \num < NUM_IRQS
+.if \num < PICO_NUM_VTABLE_IRQS
 .word \func
 .endif
 .endm
@@ -142,10 +145,11 @@ if_irq_word 76 isr_irq76
 if_irq_word 77 isr_irq77
 if_irq_word 78 isr_irq78
 if_irq_word 79 isr_irq79
-#if NUM_IRQS > 80
+#if PICO_NUM_VTABLE_IRQS > 80
 #error more IRQ entries required
 #endif
-#endif
+#endif  // #if !PICO_MINIMAL_STORED_VECTOR_TABLE
+__vectors_end:
 
 // all default exception handlers do nothing, and we can check for them being set to our
 // default values by seeing if they point to somewhere between __defaults_isrs_start and __default_isrs_end
@@ -164,24 +168,14 @@ __default_isrs_start:
     bkpt #0
 .endm
 
-// these are separated out for clarity
-decl_isr_bkpt isr_invalid
-decl_isr_bkpt isr_nmi
-decl_isr_bkpt isr_hardfault
-decl_isr_bkpt isr_svcall
-decl_isr_bkpt isr_pendsv
-decl_isr_bkpt isr_systick
-
-.global __default_isrs_end
-__default_isrs_end:
-
 .altmacro
 .macro decl_isr name
-#if !PICO_NO_STORED_VECTOR_TABLE | PICO_NO_FLASH
+#if !PICO_MINIMAL_STORED_VECTOR_TABLE | PICO_NO_FLASH
 // We declare a weak label, so user can override
 .weak \name
 #else
-// We declare a strong label, so user can't override (their version would not automatically be used)
+// We declare a strong global label, so user can't override (their version would not automatically be used)
+.global \name
 #endif
 .type \name,%function
 .thumb_func
@@ -189,8 +183,14 @@ __default_isrs_end:
 .endm
 
 .macro if_irq_decl num func
-.if \num < NUM_IRQS
+.if \num < PICO_NUM_VTABLE_IRQS
 decl_isr \func
+.elseif \num < NUM_IRQS
+// We declare a strong global label, so user can't override (their version would not automatically be used)
+.global \func
+.type \func,%function
+.thumb_func
+\func:
 .endif
 .endm
 
@@ -274,7 +274,7 @@ if_irq_decl 76 isr_irq76
 if_irq_decl 77 isr_irq77
 if_irq_decl 78 isr_irq78
 if_irq_decl 79 isr_irq79
-#if NUM_IRQS > 80
+#if PICO_NUM_VTABLE_IRQS > 80
 #error more IRQ entries required
 #endif
 
@@ -282,11 +282,29 @@ if_irq_decl 79 isr_irq79
 .global __unhandled_user_irq
 .thumb_func
 __unhandled_user_irq:
+// if we include the implementation if there could be a valid IRQ hanler in the vtable that uses it
+#if !(PICO_NO_RAM_VECTOR_TABLE && PICO_MINIMAL_STORED_VECTOR_TABLE)
     mrs  r0, ipsr
     subs r0, #16
 .global unhandled_user_irq_num_in_r0
 unhandled_user_irq_num_in_r0:
+#endif
+    // note the next instruction is a breakpoint too, however we have a 2 byte alignment hole
+    // and it is preferrable to have distinct labels, to inform the user what has happened in the debugger.
     bkpt #0
+
+decl_isr_bkpt isr_invalid
+#if !PICO_MINIMAL_STORED_VECTOR_TABLE
+// these are separated out for clarity
+decl_isr_bkpt isr_nmi
+decl_isr_bkpt isr_hardfault
+decl_isr_bkpt isr_svcall
+decl_isr_bkpt isr_pendsv
+decl_isr_bkpt isr_systick
+#endif
+
+.global __default_isrs_end
+__default_isrs_end:
 
 // ----------------------------------------------------------------------------
